+++ src/module/data/item/ability-data.mjs
@@ -1,4 +1,4 @@
-import itemDescriptionSchema from "../schema/item-description-schema.mjs";
+import itemDescriptionSchema from "./templates/item-description-schema.mjs";
 
 /**
  * Ability / Proficiency Item Data Model
