+++ src/module/data/item/templates/item-physical-schema.mjs
